requirements.txt
Flask
flask-marshmallow
marshmallow-sqlalchemy
Flask-restplus
Flask-SQLAlchemy
Flask-WTF
flask-bootstrap
flask-cors
sqlalchemy-migrate
numpy
scipy
biopython
pandas
matplotlib
sklearn
requests
gunicorn
pysam
python-dotenv
mysqlclient
click
